Started replacing sprite indices (`uint16_t`) by a class `Sprite`
However it does not build for now, and it may be more complicated than
anticipated: if the sprite instances are to be unique for a single sprite (which
is a requirement if the tiles are going to hold references to characters later
on), then the levels should own them. However there are plenty of objects
instanciating them before the levels are created, which is a problem.

+++ engine/libflatkiss-data/libflatkiss/data/loader_action_sprite_mapper.cpp
@@ -32,7 +32,7 @@ using std::string;
 using std::unordered_map;
 
 Action LoaderActionSpriteMapper::actionIdentifierToAction(
-    uint16_t action_identifier) {
+    int64_t action_identifier) {
   switch (action_identifier) {
     case 0:
       return Action::kWalkLeft;
@@ -66,9 +66,9 @@ unordered_map<int64_t, ActionSpriteMapper const> LoaderActionSpriteMapper::load(
   return index_to_mapper;
 }
 
-unordered_map<Action, uint16_t> LoaderActionSpriteMapper::loadGroup(
+unordered_map<Action, int64_t> LoaderActionSpriteMapper::loadGroup(
     int64_t group_size, ifstream& stream) {
-  unordered_map<Action, uint16_t> action_to_indices;
+  unordered_map<Action, int64_t> action_to_indices;
   for (int64_t i{0}; i < group_size; i++) {
     int64_t action_identifier{StreamReader::read(stream, 2)};
     int64_t sprite_index{StreamReader::read(stream, 2)};
